Move arguments to internal/arguments
