+++ commands/expand/expand.go
@@ -5,7 +5,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/mroth/scmpuff/commands/arguments"
+	"github.com/mroth/scmpuff/commands/internal/arguments"
 	"github.com/spf13/cobra"
 )
